+++ backend/jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/formruntime/dto/FormRecordSubmitReq.java
@@ -1,0 +1,15 @@
+package org.jeecg.modules.formruntime.dto;
+
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import lombok.Data;
+
+@Data
+@ApiModel(value = "FormRecordSubmitReq", description = "Form record submit request")
+public class FormRecordSubmitReq {
+    @ApiModelProperty(value = "form key", required = true)
+    private String formKey;
+
+    @ApiModelProperty(value = "data json", required = true)
+    private String dataJson;
+}